requirements.txt
websockets==12.0
cryptography
psycopg2-binary
python-dotenv
